Move XCTestCase async methods to their new home sweet file home.
